improve BCLK, config sequencing, demo volume
This is a best effort attempt to get BCLK working as a PLL_CLKIN
source using all the recommended best practices from the datashhet
for configuration sequencing, etc.

Note that this doesn't actually work well for BCLK, but it does do
a pretty thorough job of demonstrating how MCLK sounds better than
BCLK. That's the main point. You get substantially better audio
quality for using a 15 MHz MCLK.

I also came up with a better dac_volume and headphone_volume combo
to use for the example code to get a lower noise floor with an
approximately line level output.